requirements.txt
google-genai
python-dotenv==1.1.0
